FactoryForge_Backend/requirements.txt
asgiref==3.7.2
boto3==1.34.19
botocore==1.34.28
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
Markdown==3.5.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.10.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
wheel==0.41.2
whitenoise==6.6.0
